% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_andel_plot.R
\docType{data}
\name{.private}
\alias{.private}
\title{Shiny module providing GUI and server logic for the Andeler tab}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.private
}
\arguments{
\item{id}{Character string module namespace}
}
\value{
An shiny app ui object
}
\description{
Shiny module providing GUI and server logic for the Andeler tab
}
\keyword{datasets}
