% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_veiledning.R
\name{info_server}
\alias{info_server}
\title{Server logic}
\usage{
info_server(id, user, data)
}
\arguments{
\item{id}{Character string module namespace}

\item{user}{Shiny user object containing information about the logged-in user}

\item{data}{Data frame with the data to be used in the Rmarkdown rendering and plot}
}
\value{
A shiny app server object
}
\description{
Server logic
}
