% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_over_tid.R
\name{makeYearCountPlot}
\alias{makeYearCountPlot}
\title{Lager et søylediagram som viser antall registreringer per år}
\usage{
makeYearCountPlot(data)
}
\arguments{
\item{data}{RegData-datasettet"}
}
\value{
ggplot-objekt
}
\description{
Lager et søylediagram som viser antall registreringer per år
}
