% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_over_tid.R
\name{makeYearCountStackedPlot}
\alias{makeYearCountStackedPlot}
\title{Lager et søylediagram med stablet fyll for G20 og atypisk per år}
\usage{
makeYearCountStackedPlot(data, varChoice)
}
\arguments{
\item{data}{RegData-datasettet}

\item{varChoice}{Valgt variabel for oppsummering}
}
\value{
ggplot-objekt
}
\description{
Lager et søylediagram med stablet fyll for G20 og atypisk per år
}
