% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_over_tid.R
\name{makeYearCountTable}
\alias{makeYearCountTable}
\title{Lager en tabell som viser antall registreringer per år}
\usage{
makeYearCountTable(data, varChoice)
}
\arguments{
\item{data}{RegData-datasettet}
}
\value{
DT::datatable-objekt
}
\description{
Lager en tabell som viser antall registreringer per år
}
