% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkinsonGetRegData.R
\name{parkGetRegData}
\alias{parkGetRegData}
\title{Provides a dataframe containing data from a registry}
\usage{
parkGetRegData(datoFra = "1999-01-01", datoTil = "2099-01-01")
}
\value{
regData data frame
}
\description{
Provides a dataframe containing data from a registry
}
