% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkinsonPreProsess.R
\name{parkPreprosess}
\alias{parkPreprosess}
\title{Parkinson's Disease Preprocessing}
\usage{
parkPreprosess(bakgrunnSkjema, konsultasjonSkjema, promData)
}
\arguments{
\item{RegData}{Data frame containing the registry data}
}
\value{
A preprocessed data frame ready for analysis
}
\description{
Preprosesserer data for Parkinson
}
