% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling.R
\name{plotAvBehFordeling}
\alias{plotAvBehFordeling}
\title{Preprosessering}
\usage{
plotAvBehFordeling(data, fordelingsVariabel)
}
\arguments{
\item{data}{Data frame med data for plottene}

\item{fordelingsVariabel}{Variabelen som skal brukes for å gruppere dataene i plottene}
}
\value{
datasett med norske nivåer og verdier
}
\description{
Preprosessering
}
