% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling.R
\name{plotMedFordeling}
\alias{plotMedFordeling}
\title{Preprosessering}
\usage{
plotMedFordeling(data, fordelingsvariabel, filter_verdi = NULL)
}
\arguments{
\item{data}{datasett "licorice gargle"}
}
\value{
datasett med norske nivåer og verdier
}
\description{
Preprosessering
}
